+++ src/test/kotlin/com/petrolal/commons/web/test/TestApplication.kt
@@ -1,0 +1,6 @@
+package com.petrolal.commons.web.test
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication
+class TestApplication