+++ src/test/kotlin/com/petrolal/commons/web/test/TestEntityRepository.kt
@@ -1,0 +1,5 @@
+package com.petrolal.commons.web.test
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface TestEntityRepository : JpaRepository<TestEntity, Long>